# Refactor enocding.Chunk.Add