… CHANGELOG.md
… pkg/chunk/chunk_store_test.go
… pkg/chunk/chunk_test.go
… pkg/chunk/encoding/bigchunk_test.go
… pkg/chunk/encoding/chunk_test.go
… pkg/chunk/encoding/delta.go
… pkg/chunk/encoding/delta_test.go
… pkg/distributor/distributor_test.go
… pkg/ingester/ingester_test.go
… pkg/querier/batch/chunk_test.go
… pkg/querier/chunk_store_queryable_test.go
… pkg/querier/iterators/chunk_merge_iterator_test.go