Previously `encoding.Chunk.Add` used to return a slice of chunks including the chunk which called the `Add` method. This added a lot of unnecessary allocations to create new slices if no new chunk is created. This PR changes it to only return the new chunk if it is created, `nil` otherwise. There is no slice returned because there wont be more than 1 new chunk created for any single `Add` call.

Benchmark results for Ingester Push:
(This is with 10000 series)
```
benchmark                                        old ns/op      new ns/op      delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     2241738857     2112765505     -5.75%
BenchmarkIngesterPush/encoding=Varbit-8          2232805325     2102434453     -5.84%
BenchmarkIngesterPush/encoding=Bigchunk-8        2236451103     2084783678     -6.78%

benchmark                                        old allocs     new allocs     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     2783307        763322         -72.57%
BenchmarkIngesterPush/encoding=Varbit-8          2782899        762856         -72.59%
BenchmarkIngesterPush/encoding=Bigchunk-8        2782862        762844         -72.59%

benchmark                                        old bytes     new bytes     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     170320192     121377536     -28.74%
BenchmarkIngesterPush/encoding=Varbit-8          170238200     120911304     -28.98%
BenchmarkIngesterPush/encoding=Bigchunk-8        170210496     121198592     -28.79%
```

**Edit**
Benchmark results after reverting back to 100 series
```
benchmark                                        old ns/op     new ns/op     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     16431496      15754699      -4.12%
BenchmarkIngesterPush/encoding=Varbit-8          16342108      15671923      -4.10%
BenchmarkIngesterPush/encoding=Bigchunk-8        16409123      15624513      -4.78%

benchmark                                        old allocs     new allocs     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     32348          12139          -62.47%
BenchmarkIngesterPush/encoding=Varbit-8          32341          12138          -62.47%
BenchmarkIngesterPush/encoding=Bigchunk-8        32344          12137          -62.48%

benchmark                                        old bytes     new bytes     delta
BenchmarkIngesterPush/encoding=DoubleDelta-8     3324933       2768749       -16.73%
BenchmarkIngesterPush/encoding=Varbit-8          3341311       2764124       -17.27%
BenchmarkIngesterPush/encoding=Bigchunk-8        3338947       2769880       -17.04%
```