@tomwilkie 2019-10-03
Can we have a changelog entry making it clear we're removing the DELTA encoding please?