@bboreham 2019-10-12
Could you fix the changelog conflict please?