@codesome 2019-10-14
Done. Also merged all commits into 1.