@bboreham 2019-10-18
Personally I don't believe there are any users on Delta encoding.  The default in Cortex was DoubleDelta. I think Delta is just a left-over from ancient Prometheus history - we copied all those files when Prometheus v2 moved to different encoding - and Cortex never used it.